movies/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.11
requests==2.32.5